+++ django_apscheduler/docs/changelog.md
@@ -26,8 +26,9 @@ This changelog is used to track all major changes to django_apscheduler.
 - Switch to using `BigAutoField` for `DjangoJobExecution`'s primary keys. This should prevent running out of usable
   ID's for deployments with a very large number of job executions in the database (Resolves [#36](https://github.com/jarekwg/django-apscheduler/issues/36)).
 - Implement `DjangoJob.shutdown()` method to close database connection when scheduler is shut down.
-- **BREAKING CHANGE:** Removed `jobstores.register_events`. Calling this method is no longer necessary as the
-  `DjangoJobStore` will automatically register for events that it cares about when the scheduler is started.
+- `jobstores.register_events` has been deprecated and will be removed in a future release. Calling this method is no
+  longer necessary as the `DjangoJobStore` will automatically register for events that it cares about when the scheduler
+  is started.
 - Ensure that Django and APScheduler always use the same timezones when passing datetimes between the two.
 - Use the configured scheduler's locking mechanism to keep the creation of `DjangoJobExecution` in sync with APScheduler
   events.